Bugfix for turn off not working right when it would dip below 0

+++ include/solutions/aoc_day_6.h
@@ -42,7 +42,7 @@ class Day6BrightnessField
     public:
         Day6BrightnessField();
         ~Day6BrightnessField();
-        int sum_brightness();
+        long sum_brightness();
         void process_instruction(Day6Instruction instruction);    
 };
